+++ database/migrations/2025_03_25_165103_create_expense_member_table.php
@@ -15,8 +15,8 @@ return new class extends Migration {
             $table->foreignId('member_id')->constrained('members')->onDelete('cascade');
             $table->foreignId('expense_id')->constrained('expenses')->onDelete('cascade');
             $table->integer('ratio')->default(1);
-            $table->decimal('share', 15, 3)->default(0);
-            $table->decimal('remainder', 15, 3)->default(0);
+            $table->integer('share')->default(0);
+            $table->integer('remainder')->default(0);
             $table->timestamps();
         });
     }
